+++ extensions/instagram/src/main/java/app/morphe/extension/instagram/patches/Links.java
@@ -11,21 +11,30 @@
 
 package app.morphe.extension.instagram.patches;
 
-import android.net.Uri;
-import android.content.Context;
 import android.content.Intent;
+import android.net.Uri;
+
+import java.io.IOException;
 import java.net.URI;
 import java.util.List;
-import java.io.IOException;
 
+import app.morphe.extension.instagram.entity.Entity;
+import app.morphe.extension.instagram.settings.SettingsStatus;
 import app.morphe.extension.instagram.utils.Pref;
 import app.morphe.extension.shared.Logger;
 import app.morphe.extension.shared.Utils;
-import app.morphe.extension.instagram.settings.SettingsStatus;
-import app.morphe.extension.instagram.entity.Entity;
 
+@SuppressWarnings("unused")
 public class Links {
-    private static boolean DISABLE_ANALYTICS,VIEW_STORIES_ANONYMOUSLY,VIEW_LIVE_ANONYMOUSLY,DISABLE_STORIES,DISABLE_EXPLORE,DISABLE_COMMENTS,DISABLE_DISCOVER_PEOPLE;
+    private static final boolean DISABLE_ANALYTICS;
+    private static final boolean VIEW_STORIES_ANONYMOUSLY;
+    private static final boolean VIEW_LIVE_ANONYMOUSLY;
+    private static final boolean DISABLE_STORIES;
+    private static final boolean DISABLE_EXPLORE;
+    private static final boolean DISABLE_COMMENTS;
+    private static final boolean DISABLE_DISCOVER_PEOPLE;
+    private static final boolean DISABLE_ADS;
+
     static {
         DISABLE_ANALYTICS = Pref.disableAnalytics() && SettingsStatus.disableAnalytics;
         VIEW_STORIES_ANONYMOUSLY = Pref.viewStoriesAnonymously() && SettingsStatus.viewStoriesAnonymously;
@@ -34,6 +43,7 @@ public class Links {
         DISABLE_EXPLORE = Pref.disableExplore() && SettingsStatus.disableExplore;
         DISABLE_COMMENTS = Pref.disableComments() && SettingsStatus.disableComments;
         DISABLE_DISCOVER_PEOPLE = Pref.disableDiscoverPeople() && SettingsStatus.disableDiscoverPeople;
+        DISABLE_ADS = Pref.disableAds() && SettingsStatus.disableAds;
     }
 
 
@@ -95,6 +105,11 @@ public class Links {
                     shouldBlockUri = DISABLE_COMMENTS;
                 } else if (path.contains("/discover/ayml/")) {
                     shouldBlockUri = DISABLE_DISCOVER_PEOPLE;
+                } else if (path.contains("profile_ads/get_profile_ads/")
+                        || path.contains("/async_ads/")
+                        || path.contains("/feed/injected_reels_media/")
+                        || path.contains("/api/v1/ads/graphql/")) {
+                    shouldBlockUri = DISABLE_ADS;
                 }
 
             }