+++ extensions/instagram/src/main/java/app/morphe/extension/instagram/settings/SettingsStatus.java
@@ -80,19 +80,27 @@ public class SettingsStatus {
     public static void hideStoriesTray() {
         hideStoriesTray = true;
     }
+    public static boolean limitFollowingFeed = false;
+    public static void limitFollowingFeed() {
+        limitFollowingFeed = true;
+    }
+    public static boolean hideGroupCreationOnSharesheet = false;
+    public static void hideGroupCreationOnSharesheet() {
+        hideGroupCreationOnSharesheet = true;
+    }
     public static boolean distractionFreeSection() {
-        return (disableStories || disableExplore || disableComments || hideStoriesTray);
+        return (disableStories || disableExplore || disableComments || hideStoriesTray || limitFollowingFeed || hideGroupCreationOnSharesheet);
     }
 
     //Misc section.
     public static boolean disableAnalytics = false;
-    public static void disableAnalytics() {disableAnalytics = true;}
+    public static void disableAnalytics() { disableAnalytics = true; }
     public static boolean disableDiscoverPeople = false;
     public static void disableDiscoverPeople() {
         disableDiscoverPeople = true;
     }
     public static boolean followBackIndicator = false;
-    public static void followBackIndicator() {followBackIndicator = true;}
+    public static void followBackIndicator() { followBackIndicator = true; }
     public static boolean viewStoryMentions = false;
     public static void viewStoryMentions() {
         viewStoryMentions = true;
@@ -109,11 +117,22 @@ public class SettingsStatus {
     public static void unlimitedReplaysOnEphemeralMedia() {
         unlimitedReplaysOnEphemeralMedia = true;
     }
-    public static boolean miscSection() {return (unlimitedReplaysOnEphemeralMedia || customiseStoryTimestamp || disableAnalytics || disableDiscoverPeople || followBackIndicator || viewStoryMentions || disableStoryFlipping);}
+    public static boolean improveImageViewing = false;
+    public static void improveImageViewing() {
+        improveImageViewing = true;
+    }
+    public static boolean hideReshareButton = false;
+    public static void hideReshareButton() {
+        hideReshareButton = true;
+    }
+    public static boolean miscSection() {return (improveImageViewing || unlimitedReplaysOnEphemeralMedia || customiseStoryTimestamp || disableAnalytics || disableDiscoverPeople || followBackIndicator || viewStoryMentions || disableStoryFlipping || hideReshareButton);}
 
     //Download section.
     public static boolean downloadMedia = false;
     public static void downloadMedia() {downloadMedia = true;}
+
+    public static boolean hideNavigationButtons = false;
+    public static void hideNavigationButtons() { hideNavigationButtons = true; }
 
     public static void load() {
     }