+++ extensions/twitter/src/main/java/app/morphe/extension/twitter/settings/ActivityHook.java
@@ -122,10 +122,10 @@ public class ActivityHook {
             toolbarText = "piko_title_logging";
         }else if (activity_name.equals(Settings.READER_MODE_KEY)) {
             toolbarText = "piko_title_native_reader_mode";
-        }else if (activity_name.equals(Settings.READER_MODE_KEY)) {
-            toolbarText = "piko_title_native_reader_mode";
         }else if (activity_name.equals(Settings.CHANGE_APP_ICON)) {
             toolbarText = "piko_pref_customisation_change_app_icon";
+        }else if (activity_name.equals(Settings.EXPORT_LOGIN_TOKEN)) {
+            toolbarText = "piko_pref_export_login_token";
         }
         return Utils.getResourceString(toolbarText);
     }