+++ patches/src/main/kotlin/app/crimera/patches/instagram/utils/Constants.kt
@@ -20,13 +20,20 @@ object Constants {
             name = "Instagram",
             packageName = "com.instagram.android",
             apkFileType = ApkFileType.APKM,
-            appIconColor = 0x000000,
+            appIconColor = 0xFC483C,
             targets =
                 listOf(
                     // Stable
-                    AppTarget(version = "422.0.0.44.64"),
+                    AppTarget(
+                        version = "423.0.0.47.66",
+                        description = "Instagram Stable version (all archs)",
+                    ),
                     // Alpha
-                    AppTarget(version = "424.0.0.0.39"),
+                    AppTarget(
+                        version = "425.0.0.0.0",
+                        description = "Instagram Alpha version (arm64-v8a only)",
+                        isExperimental = true,
+                    ),
                 ),
         )